Python/GamingLights/requirements.txt
pyserial
bitstring